Linting 38

+++ Source/Components/GAMs/SimulinkWrapperGAM/SimulinkClasses.h
@@ -152,7 +152,7 @@ class SimulinkParameter : public SimulinkDataI {
     
 public:
 
-    uint32 offset;      //!< Parameter offset (used if the parameter is part of a structure).
+    uint64 offset;      //!< Parameter offset (used if the parameter is part of a structure).
 
     /**
      * @brief   Updates the value of a parameter inside the model
